backend/requirements.txt
fastapi==0.116.1
google-generativeai==0.8.5
python-dotenv==1.1.1
python-jose==3.5.0
supabase==2.18.1
uvicorn==0.35.0
